pyproject.toml
[project]
name = "flux-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.20.0.post0",
    "loguru>=0.7.3",
    "pydantic-settings>=2.9.1",
    "redis>=6.2.0",
    "runware>=0.4.11",
    "websockets>=15.0.1",
]
